requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.0b2
Django==4.2.15
django-bootstrap4==24.3
django-pictures==1.3.3
django-stdimage==6.0.2
gunicorn==23.0.0
importlib-metadata==8.4.0
importlib-resources==6.4.4
jaraco.context==6.0.1
jaraco.functools==4.0.2
jaraco.text==4.0.0
more-itertools==10.4.0
mysql==0.0.3
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyMySQL==1.1.1
requests==2.32.3
setuptools==73.0.1
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.1
wheel==0.44.0
whitenoise==6.7.0
zipp==3.20.0